refactor(use_case): rename server_graphql.py → mcp_server.py, trim __init__
- server_graphql.py → mcp_server.py. The "_graphql" suffix was a triple
  anachronism: the classic server variant was removed, the file no longer
  uses graphql-core types, and the name overlapped with the sibling
  pydantic_resolve/graphql/ package. Updates the 3 imports + 1 docstring
  that referenced the old path.
- Drop UseCaseResources from the package __all__. External code obtains
  it via UseCaseManager.get_app(); nobody constructs it directly. Keeping
  it in the public surface suggested it was meant for direct instantiation.

739 tests pass.
